feat: hardening de seguranca — sessao HttpOnly, rate limit, money.js, recuperar senha

+++ src/migrations/001-create-users-table.js
@@ -15,7 +15,7 @@ export async function up(queryInterface, Sequelize) {
       allowNull: true
     },
     saldo: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.DECIMAL(14, 2),
       allowNull: true,
       defaultValue: 0
     },
