+++ src/migrations/003-create-pagamentos-table.js
@@ -31,7 +31,7 @@ export async function up(queryInterface, Sequelize) {
       allowNull: true
     },
     valor: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.DECIMAL(14, 2),
       allowNull: true,
       defaultValue: 0
     },
